Revert "fix(core): retry malformed streamed responses (#6754)"
This reverts commit 43d34f739d6a64af66981b478cf9da667d67a4ed.
