+++ packages/core/src/core/openaiContentGenerator/converter.test.ts
@@ -96,18 +96,6 @@ describe('OpenAIContentConverter', () => {
   }
 
   describe('stream-local parser state', () => {
-    const streamChunk = (
-      id: string,
-      delta: Record<string, unknown>,
-      finishReason: string | null = null,
-    ) =>
-      ({
-        id,
-        created: 1,
-        model: 'test',
-        choices: [{ index: 0, delta, finish_reason: finishReason }],
-      }) as unknown as OpenAI.Chat.ChatCompletionChunk;
-
     it('creates fresh parser instances', () => {
       const ctx1 = new StreamingToolCallParser();
       const ctx2 = new StreamingToolCallParser();
@@ -362,283 +350,6 @@ describe('OpenAIContentConverter', () => {
       expect(fn?.name).toBe('list_sessions');
       expect(fn?.args).toEqual({});
       expect(fn?.id).toBe('call_noargs');
-    });
-
-    it('discards a response whose tool call never provides a function name', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-
-      expect(
-        converter.convertOpenAIChunkToGemini(
-          streamChunk('malformed-tool-call', {
-            reasoning_content: 'failed reasoning',
-            content: 'failed visible content',
-            tool_calls: [
-              {
-                index: 0,
-                id: 'call_without_name',
-                type: 'function',
-                function: { arguments: '' },
-              },
-            ],
-          }),
-          stream,
-        ).candidates?.[0]?.content?.parts,
-      ).toEqual([]);
-      const result = converter.convertOpenAIChunkToGemini(
-        streamChunk('malformed-tool-call-finish', {}, 'stop'),
-        stream,
-      );
-      expect(result.candidates?.[0]?.content?.parts).toEqual([]);
-      expect(result.candidates?.[0]?.finishReason).toBeUndefined();
-    });
-
-    it('preserves literal thinking tags when no nameless tool call appears', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-
-      expect(
-        converter.convertOpenAIChunkToGemini(
-          streamChunk('literal-think', {
-            content: 'Use <think>literal</think> text.',
-          }),
-          stream,
-        ).candidates?.[0]?.content?.parts,
-      ).toEqual([{ text: 'Use <think>literal</think> text.' }]);
-
-      const result = converter.convertOpenAIChunkToGemini(
-        streamChunk('literal-think-finish', {}, 'stop'),
-        stream,
-      );
-
-      expect(result.candidates?.[0]?.content?.parts).toEqual([]);
-      expect(result.candidates?.[0]?.finishReason).toBe(FinishReason.STOP);
-    });
-
-    it('preserves inline thinking-tag references after structured reasoning', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-
-      const reasoning = converter.convertOpenAIChunkToGemini(
-        streamChunk('literal-think-reasoning', {
-          reasoning_content: 'The user is asking about response formats.',
-        }),
-        stream,
-      );
-      const content = converter.convertOpenAIChunkToGemini(
-        streamChunk('literal-think-content', {
-          content:
-            'In Qwen output, the `<think>` tag wraps hidden reasoning text.',
-        }),
-        stream,
-      );
-      const finish = converter.convertOpenAIChunkToGemini(
-        streamChunk('literal-think-finish', {}, 'stop'),
-        stream,
-      );
-
-      expect(reasoning.candidates?.[0]?.content?.parts).toEqual([
-        {
-          thought: true,
-          text: 'The user is asking about response formats.',
-        },
-      ]);
-      expect(content.candidates?.[0]?.content?.parts).toEqual([
-        {
-          text: 'In Qwen output, the `<think>` tag wraps hidden reasoning text.',
-        },
-      ]);
-      expect(finish.candidates?.[0]?.finishReason).toBe(FinishReason.STOP);
-    });
-
-    it('drops buffered thinking tags when a later tool call has no name', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-
-      converter.convertOpenAIChunkToGemini(
-        streamChunk('late-think', {
-          content: 'late <think>payload</think>',
-        }),
-        stream,
-      );
-      const malformed = converter.convertOpenAIChunkToGemini(
-        streamChunk('late-tool-call', {
-          tool_calls: [
-            {
-              index: 0,
-              id: 'call_without_name',
-              type: 'function',
-              function: { arguments: '' },
-            },
-          ],
-        }),
-        stream,
-      );
-      const result = converter.convertOpenAIChunkToGemini(
-        streamChunk('late-finish', {}, 'stop'),
-        stream,
-      );
-
-      expect(malformed.candidates?.[0]?.content?.parts).toEqual([]);
-      expect(result.candidates?.[0]?.content?.parts).toEqual([]);
-      expect(result.candidates?.[0]?.finishReason).toBeUndefined();
-    });
-
-    it('rejects raw thinking tags that leak after structured reasoning output', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-      const reasoning = converter.convertOpenAIChunkToGemini(
-        streamChunk('structured-reasoning', {
-          reasoning_content: 'Let me check<think>',
-        }),
-        stream,
-      );
-      const leaked = converter.convertOpenAIChunkToGemini(
-        streamChunk('leaked-visible-thinking', {
-          content:
-            ' file read the presubmit report.\n</think>\n<think>\n\n</think>\n\n',
-        }),
-        stream,
-      );
-      expect(reasoning.candidates?.[0]?.content?.parts).toEqual([]);
-      expect(leaked.candidates?.[0]?.content?.parts).toEqual([]);
-      expect(() =>
-        converter.convertOpenAIChunkToGemini(
-          streamChunk('leaked-visible-thinking-finish', {}, 'stop'),
-          stream,
-        ),
-      ).toThrowError(expect.objectContaining({ type: 'PROTOCOL_TAG_LEAK' }));
-    });
-
-    it.each([
-      ['closing tag', '</thi', 'nk> leaked visible reasoning'],
-      [
-        'mixed-case tag with whitespace',
-        '<Thinking ',
-        '>leaked visible reasoning</Thinking >',
-      ],
-    ])(
-      'rejects a split %s across streamed deltas',
-      (_name, firstDelta, secondDelta) => {
-        const stream = withStreamParser(new StreamingToolCallParser());
-        converter.convertOpenAIChunkToGemini(
-          streamChunk('structured-reasoning-before-split-leak', {
-            reasoning_content: 'hidden reasoning',
-          }),
-          stream,
-        );
-
-        const firstLeak = converter.convertOpenAIChunkToGemini(
-          streamChunk('split-leak-start', { content: firstDelta }),
-          stream,
-        );
-        const secondLeak = converter.convertOpenAIChunkToGemini(
-          streamChunk('split-leak-end', { content: secondDelta }),
-          stream,
-        );
-
-        expect(firstLeak.candidates?.[0]?.content?.parts).toEqual([]);
-        expect(secondLeak.candidates?.[0]?.content?.parts).toEqual([]);
-        expect(() =>
-          converter.convertOpenAIChunkToGemini(
-            streamChunk('split-leak-finish', {}, 'stop'),
-            stream,
-          ),
-        ).toThrowError(expect.objectContaining({ type: 'PROTOCOL_TAG_LEAK' }));
-      },
-    );
-
-    it('rejects an incomplete thinking tag prefix when the stream finishes', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-      converter.convertOpenAIChunkToGemini(
-        streamChunk('structured-reasoning-before-incomplete-leak', {
-          reasoning_content: 'hidden reasoning',
-        }),
-        stream,
-      );
-      converter.convertOpenAIChunkToGemini(
-        streamChunk('incomplete-leak', {
-          content: 'leaked visible reasoning </thi',
-        }),
-        stream,
-      );
-
-      expect(() =>
-        converter.convertOpenAIChunkToGemini(
-          streamChunk('incomplete-leak-finish', {}, 'stop'),
-          stream,
-        ),
-      ).toThrowError(expect.objectContaining({ type: 'PROTOCOL_TAG_LEAK' }));
-    });
-
-    it('marks a completed structured reasoning tag leak explicitly', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-
-      expect(() =>
-        converter.convertOpenAIChunkToGemini(
-          streamChunk(
-            'explicit-protocol-tag-leak',
-            {
-              reasoning_content: 'hidden reasoning',
-              content: '</think> leaked visible reasoning',
-            },
-            'stop',
-          ),
-          stream,
-        ),
-      ).toThrowError(
-        expect.objectContaining({
-          name: 'InvalidStreamError',
-          type: 'PROTOCOL_TAG_LEAK',
-        }),
-      );
-    });
-
-    it('rejects raw thinking tags when they share a chunk with structured reasoning', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-
-      expect(() =>
-        converter.convertOpenAIChunkToGemini(
-          streamChunk(
-            'same-chunk-structured-reasoning-leak',
-            {
-              reasoning_content: 'hidden reasoning',
-              content: '<think>leaked reasoning</think>visible',
-            },
-            'stop',
-          ),
-          stream,
-        ),
-      ).toThrowError(expect.objectContaining({ type: 'PROTOCOL_TAG_LEAK' }));
-    });
-
-    it('rejects tag leaks before emitting their tool calls', () => {
-      const stream = withStreamParser(new StreamingToolCallParser());
-      const reasoning = converter.convertOpenAIChunkToGemini(
-        streamChunk('structured-reasoning-before-tool-leak', {
-          reasoning_content: 'hidden reasoning',
-        }),
-        stream,
-      );
-
-      expect(reasoning.candidates?.[0]?.content?.parts).toEqual([
-        { thought: true, text: 'hidden reasoning' },
-      ]);
-      expect(() =>
-        converter.convertOpenAIChunkToGemini(
-          streamChunk(
-            'structured-reasoning-leak-with-tool-call',
-            {
-              content: '</think> leaked visible reasoning',
-              tool_calls: [
-                {
-                  index: 0,
-                  id: 'call_named',
-                  type: 'function',
-                  function: { name: 'run_shell_command', arguments: '{}' },
-                },
-              ],
-            },
-            'tool_calls',
-          ),
-          stream,
-        ),
-      ).toThrowError(expect.objectContaining({ type: 'PROTOCOL_TAG_LEAK' }));
     });
   });
